cc-week07/day03-hw-vehicles/src/main/java/Driver.java
public class Driver {

    String name;
    IDriveable vehicle;

    public Driver(String name, IDriveable vehicle) {
        this.name = name;
        this.vehicle = vehicle;
    }

    public String getName() {
        return this.name;
    }

    public IDriveable getVehicle() {
        return this.vehicle;
    }

    public int driveDistance(int distance) {
        return this.vehicle.driveDistance(distance);
    }

    public void setVehicle(IDriveable vehicle) {
        this.vehicle = vehicle;
    }
}
